Avoid promise constructor anti-pattern in readme
I see no reason to wrap the promise during this example. At least it also anti-pattern as described here: https://github.com/petkaantonov/bluebird/wiki/Promise-anti-patterns#the-deferred-anti-pattern

The only difference between the proposed code, and the existing code is, that the old one returns an `Promise<boolean>` and the new one returns a `Promise<void>`, which makes no difference in this particular case.